requirements.txt
fastapi
python-multipart
uvicorn
PyPDF2
langchain-community
langchain-core
langchain-google-genai
langgraph
pydantic
python-dotenv
pandas
openpyxl
langchain-redis
redis
celery
psycopg2-binary